LPS-181298 Rename images according with existing patterns and delete unused files
