+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/assets/images/account_placeholder.png
