+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/assets/images/app_placeholder.png
