+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/assets/images/no_app.png
