+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/components/DashboardTable/DashboardEmptyTable.tsx
@@ -1,4 +1,4 @@
-import noApp from '../../assets/images/no-app.png';
+import noApp from '../../assets/images/no_app.png';
 
 import './DashboardEmptyTable.scss';
