+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherGatePage/PublisherGatePage.tsx
@@ -1,7 +1,7 @@
 import ClayButton from '@clayui/button';
 
 import cash from '../../assets/images/cash.svg';
-import cloudUpload from '../../assets/images/cloud-upload.svg';
+import cloudUpload from '../../assets/images/cloud_upload.svg';
 import {GateCard} from '../../components/Card/GateCard';
 import {Footer} from '../../components/Footer/Footer';
 import {Header} from '../../components/Header/Header';
